% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{checkerboard_partitioning}
\alias{checkerboard_partitioning}
\title{Partition into checkerboard (spatial folding with k=4)}
\usage{
checkerboard_partitioning(occ, occ.z, bg, bg.z, modelstack)
}
\arguments{
\item{modelstack}{}
}
\value{

}
\description{
Partition into checkerboard (spatial folding with k=4)
}
