% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{create_background_points_from_survey}
\alias{create_background_points_from_survey}
\title{Create background points from survey points (on the basis of bias file)}
\usage{
create_background_points_from_survey(modelstack, pts, nb)
}
\arguments{
\item{pts}{}
}
\value{

}
\description{
Create background points from survey points (on the basis of bias file)
}
