% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{define_points_along_survey_transects}
\alias{define_points_along_survey_transects}
\title{Define equally spaced points along survey transects}
\usage{
define_points_along_survey_transects(trans, length)
}
\arguments{
\item{length}{}
}
\value{

}
\description{
Define equally spaced points along survey transects
}
