% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate niche model with user partition}
\usage{
evaluate_model(occ, modelstack, bg, args, user.grp, type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Evaluate niche model with user partition
}
