% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{extract_bathy_at_point}
\alias{extract_bathy_at_point}
\title{Extract bathymetric data at point (presence or pseudo-absence point)}
\usage{
extract_bathy_at_point(modelstack, type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Extract bathymetric data at point (presence or pseudo-absence point)
}
