% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{get_predictors_contributions}
\alias{get_predictors_contributions}
\title{Get predictor contributions}
\usage{
get_predictors_contributions(modelstack, occ, bg, type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Get predictor contributions
}
