% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{get_predictors_importance}
\alias{get_predictors_importance}
\title{Get predictor importance}
\usage{
get_predictors_importance(modelstack, occ, bg, type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Get predictor importance
}
