% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{get_random_points}
\alias{get_random_points}
\title{Get a given number of random points}
\usage{
get_random_points(modelstack, ext, occ, nb)
}
\arguments{
\item{nb}{}
}
\value{

}
\description{
Get a given number of random points
}
