% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{read_bathy_data}
\alias{read_bathy_data}
\title{Read rasters of bathymetrically derived predictors}
\usage{
read_bathy_data(var)
}
\arguments{
\item{var}{}
}
\value{

}
\description{
Read rasters of bathymetrically derived predictors
}
