% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{remove_duplicates_occ}
\alias{remove_duplicates_occ}
\title{Remove duplicate occurrences (ie those that share a grid cell in the predictor variable raster)}
\usage{
remove_duplicates_occ(modelstack, occ)
}
\arguments{
\item{occ}{}
}
\value{

}
\description{
Remove duplicate occurrences (ie those that share a grid cell in the predictor variable raster)
}
