% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_data.R
\name{train_test_partition}
\alias{train_test_partition}
\title{Partition occurrences into train and test datasets}
\usage{
train_test_partition(occ)
}
\arguments{
\item{occ}{}
}
\value{

}
\description{
Partition occurrences into train and test datasets
}
